Confidence interval for G-index difference in a 2-group design

Computes adjusted Wald confidence intervals for the G-index of agreement within each group and the difference of G-indices.

Usage

ci.agree2(alpha, n1, f1, n2, f2, r)

Arguments

alpha

alpha level for 1-alpha confidence

n1

sample size (objects) in group 1

f1

number of objects rated in agreement in group 1

n2

sample size (objects) in group 2

f2

number of objects rated in agreement in group 2

r

number of rating categories

Value

Returns a 3-row matrix. The rows are:

Row 1: G-index for group 1

Row 2: G-index for group 2

Row 3: G-index difference

The columns are:

Estimate - maximum likelihood estimate of G-index and difference

SE - standard error

LL - lower limit of adjusted Wald confidence interval

UL - upper limit of adjusted Wald confidence interval

References

Bonett DG (2022). “Statistical inference for G-indices of agreement.” Journal of Educational and Behavioral Statistics, 47(4), 438–458. doi:10.3102/10769986221088561 .

Examples

ci.agree2(.05, 75, 70, 60, 45, 2)
#>          Estimate         SE        LL        UL
#> G1      0.8666667 0.02880329 0.6974555 0.9481141
#> G2      0.5000000 0.05590170 0.2523379 0.6851621
#> G1 - G2 0.3666667 0.06288585 0.1117076 0.6088621

# Should return:
#          Estimate         SE        LL        UL
# G1      0.8666667 0.02880329 0.6974555 0.9481141
# G2      0.5000000 0.05590170 0.2523379 0.6851621
# G1 - G2 0.3666667 0.06288585 0.1117076 0.6088621