Bayesian credible interval for a single proportion

Computes a Bayesian credible interval for a single proportion using the mean and standard deviation of a prior Beta distribution along with sample information. The mean and standard deviation of the posterior Beta distribution are also reported. For a noninformative prior, set the prior mean to .5 and the prior standard deviation to 1/sqrt(12) (which corresponds to a Beta(1,1) distribution). The prior variance must be less than m(1 - m) where m is the prior mean.

ci.bayes.prop1(alpha, prior.mean, prior.sd, f, n)

Arguments

alpha

alpha level for 1-alpha credibility interval

prior.mean

mean of prior Beta distribution

prior.sd

standard deviation of prior Beta distribution

f

number of participants who have the attribute

n

sample size

Value

Returns a 1-row matrix. The columns are:

Posterior mean - posterior mean of Beta distributoin

Posterior SD - posterior standard deviation of Beta distributoin

LL - lower limit of the credible interval

UL - upper limit of the credible interval

References

Gelman A, B. CJ, Stern HS, Rubin DB (2004). Bayesian Data Analysis, 2nd edition. Chapman & Hall.

Examples

ci.bayes.prop1(.05, .4, .1, 12, 100)
#>      Posterior mean Posterior SD        LL        UL
#> [1,]      0.1723577   0.03419454 0.1111747 0.2436185

# Should return:
#      Posterior mean Posterior SD       LL        UL
# [1,]           0.15   0.03273268  0.09218 0.2188484