Confidence interval for a biserial-phi coefficient

Computes a confidence interval for a biserial-phi coefficient using a transformation of a confidence interval for an odds ratio with .5 added to each cell frequency. This measure of association assumes the group variable is naturally dichotomous and the response variable is artificially dichotomous.

ci.biphi(alpha, f1, f2, n1, n2)

Arguments

alpha
alpha level for 1-alpha confidence

f1
number of participants in group 1 who have the attribute

f2
number of participants in group 2 who have the attribute

n1
sample size of group 1

n2
sample size of group 2

Value

Returns a 1-row matrix. The columns are:

Estimate - estimate of phi coefficient

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Ulrich R, Wirtz M (2004). “On the correlation of a naturally and an artificially dichotomized variable.” British Journal of Mathematical and Statistical Psychology, 57(2), 235--251. ISSN 00071102, doi: 10.1348/0007110042307203 , https://doi.wiley.com/10.1348/0007110042307203.

Examples

ci.biphi(.05, 46, 15, 100, 100)

#>       Estimate         SE        LL       UL
#> [1,] 0.4145733 0.07551281 0.2508866 0.546141

# Should return:
#       Estimate         SE        LL       UL
# [1,] 0.4145733 0.07551281 0.2508866 0.546141