Confidence interval for Cramer's V

Computes a confidence interval for a population Cramer's V coefficient of nominal association for an r x s contingency table. The confidence interval is based on a noncentral chi-square distribution, and an approximate standard error is recovered from the confidence interval.

Usage

ci.cramer(alpha, chisqr, r, c, n)

Arguments

alpha

alpha value for 1-alpha confidence

chisqr

Pearson chi-square test statistic of independence

r

number of rows in contingency table

c

number of columns in contingency table

n

sample size

Value

Returns a 1-row matrix. The columns are:

Estimate - estimate of Cramer's V

SE - recovered standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Smithson M (2003). Confidence Intervals. Sage.

Examples

ci.cramer(.05, 19.21, 2, 3, 200)
#>  Estimate     SE     LL     UL
#>    0.3099 0.0718 0.1601 0.4417

# Should return:
# Estimate     SE     LL     UL
#   0.3099 0.0718 0.1601 0.4417