Bonferroni confidence intervals for all pairwise proportion differences in a between-subjects design

Computes adjusted Wald confidence intervals for all pairwise differences of proportions in a between-subjects design with a Bonferroni adjusted alpha level.

ci.pairs.prop.bs(alpha, f, n)

Arguments

alpha
alpha level for simultaneous 1-alpha confidence

f
vector of frequency counts of participants with attribute

n
vector of sample sizes

Value

Returns a 1-row matrix. The columns are:

Estimate - adjusted estimate of proportion difference

SE - adjusted standard error

z - z test statistic

p - p-value

LL - lower limit of the adjusted Wald confidence interval

UL - upper limit of the adjusted Wald confidence interval

References

Agresti A, Caffo B (2000). “Simple and effective confidence intervals for proportions and differences of proportions result from adding two successes and two failures.” The American Statistician, 54(4), 280. ISSN 00031305, doi: 10.2307/2685779 , https://www.jstor.org/stable/2685779?origin=crossref.

Examples

f <- c(111, 161, 132)
n <- c(200, 200, 200)
ci.pairs.prop.bs(.05, f, n)

#>        Estimate         SE         z            p          LL          UL
#>  1 2 -0.2475248 0.04482323 -5.522243 3.346989e-08 -0.35483065 -0.14021885
#>  1 3 -0.1039604 0.04833562 -2.150803 3.149174e-02 -0.21967489  0.01175409
#>  2 3  0.1435644 0.04358401  3.293968 9.878366e-04  0.03922511  0.24790360

# Should return:
#        Estimate         SE         z            p          LL          UL
# 1 2  -0.2475248 0.04482323 -5.522243 3.346989e-08 -0.35483065 -0.14021885
# 1 3  -0.1039604 0.04833562 -2.150803 3.149174e-02 -0.21967489  0.01175409
# 2 3   0.1435644 0.04358401  3.293968 9.878366e-04  0.03922511  0.24790360