Confidence interval for a 2-group proportion difference

Computes an adjusted Wald confidence interval for a population proportion difference in a 2-group design.

Usage

ci.prop2(alpha, f1, f2, n1, n2)

Arguments

alpha

alpha level for 1-alpha confidence

f1

number of participants in group 1 who have the attribute

f2

number of participants in group 2 who have the attribute

n1

sample size for group 1

n2

sample size for group 2

Value

Returns a 1-row matrix. The columns are:

Estimate - adjusted estimate of proportion difference

SE - adjusted standard error

LL - lower limit of the adjusted Wald confidence interval

UL - upper limit of the adjusted Wald confidence interval

References

Agresti A, Caffo B (2000). “Simple and effective confidence intervals for proportions and differences of proportions result from adding two successes and two failures.” The American Statistician, 54(4), 280-288. ISSN 00031305, doi:10.2307/2685779 .

Examples

ci.prop2(.05, 35, 21, 150, 150)
#>    Estimate         SE          LL        UL
#>  0.09210526 0.04476077 0.004375769 0.1798348

# Should return:
#   Estimate         SE          LL        UL
# 0.09210526 0.04476077 0.004375769 0.1798348