Confidence interval for squared multiple correlation

Computes an approximate confidence interval for a population squared multiple correlation in a linear model with random predictor variables. This function uses the scaled central F approximation method.

ci.rsqr(alpha, r2, s, n)

Arguments

alpha

alpha value for 1-alpha confidence

r2

estimated unadjusted squared multiple correlation

s

number of predictor variables

n

sample size

Value

Returns a 1-row matrix. The columns are:

R-squared - estimate of unadjusted R-squared

adj R-squared - bias adjusted R-squared estimate

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Helland IS (1987). “On the interpretation and use of R2 in regression analysis.” Biometrics, 43(1), 61-69. doi: 10.2307/2531949 , https://doi.org/10.2307/2531949.

Examples

ci.rsqr(.05, .241, 3, 116)
#>      R-squared adj R-squared         LL        UL
#> [1,]     0.241     0.2206696 0.09819599 0.3628798

# Should return:
#        R-squared    adj R-squared          LL        UL
# [1,]       0.241        0.2206696  0.09819599 0.3628798