Confidence interval for a 2-group Spearman correlation difference

Computes a confidence interval for a difference in population Spearman correlations in a 2-group design.

ci.spear2(alpha, cor1, cor2, n1, n2)

Arguments

alpha
alpha level for 1-alpha confidence

cor1
sample Spearman correlation in group 1

cor2
sample Spearman correlation in group 2

n1
sample size in group 1

n2
sample size in group 2

Value

Returns a 1-row matrix. The columns are:

Estimate - estimated correlation difference

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG, Wright TA (2000). “Sample size requirements for estimating Pearson, Kendall and Spearman correlations.” Psychometrika, 65(1), 23--28. ISSN 0033-3123, doi: 10.1007/BF02294183 , https://link.springer.com/10.1007/BF02294183.

Zou GY (2007). “Toward using confidence intervals to compare correlations.” Psychological Methods, 12(4), 399--413. ISSN 1939-1463, doi: 10.1037/1082-989X.12.4.399 , https://doi.apa.org/getdoi.cfm?doi=10.1037/1082-989X.12.4.399.

Examples

ci.spear2(.05, .54, .48, 180, 200)

#>      Estimate         LL        UL
#> [1,]     0.06 -0.1003977 0.2185085

# Should return:
#      Estimate         LL        UL
# [1,]     0.06 -0.1003977 0.2185085