SEM fit indices

Computes the normed fit index (NFI), adjusted normed fit index (adj NFI), comparative fit index (CFI), Tucker-Lewis fit index (TLI), and root mean square error of approximation index (RMSEA). Of the first four indices, the adj NFI index is recommended because it has smaller sampling variability than CFI and TLI and less negative bias than NFI.

Usage

fitindices(chi1, df1, chi2, df2, n)

Arguments

chi1

chi-square test statistic for full model

df1

degrees of freedom for full model

chi2

chi-square test statistic for reduced model

df2

degrees of freedom for reduced model

n

sample size

Value

Returns NFI, adj NFI, CFI, TLI, and RMSEA

Examples

fitindices(14.21, 10, 258.43, 20, 300)
#>        NFI   adj NFI       CFI       TLI      RMSEA
#>  0.9450141 0.9837093 0.9823428 0.9646857 0.03746109

# Should return:
#        NFI   adj NFI       CFI       TLI      RMSEA
#  0.9450141 0.9837093 0.9823428 0.9646857 0.03746109