Sample size for a 2-group Spearman correlation difference confidence interval

Computes the sample size required to estimate a difference in population Spearman correlations with desired confidence interval precision in a 2-group design. Set the correlation planning values to the smallest absolute values within their plausible ranges for a conservatively large sample size.

Usage

size.ci.spear2(alpha, cor1, cor2, w)

Arguments

alpha

alpha level for 1-alpha confidence

cor1

group 1 Spearman correlation planning value

cor2

group 2 Spearman correlation planning value

w

desired confidence interval width

Value

Returns the required sample size

References

Bonett DG, Wright TA (2000). “Sample size requirements for estimating Pearson, Kendall and Spearman correlations.” Psychometrika, 65(1), 23–28. ISSN 0033-3123, doi:10.1007/BF02294183 .

Examples

size.ci.spear2(.05, .8, .5, .2)
#>  Sample size per group
#>                    314

# Should return:
# Sample size per group
#                   314