Confidence interval for a 2-group standardized mean difference with stratified sampling

Computes confidence intervals for a population standardized mean difference in a 2-group nonexperimental design with stratified random sampling using a square root weighted variance standardizer or single group standard deviation standardizer. Equality of variances is not assumed.

ci.stdmean.strat(alpha, m1, m2, sd1, sd2, n1, n2, p1)

Arguments

alpha
alpha level for 1-alpha confidence

m1
sample mean in group 1

m2
sample mean in group 2

sd1
sample standard deviation in group 1

sd2
sample standard deviation in group 2

n1
sample size in group 1

n2
sample size in group 2

p1
proportion of total population in subpopulation 1

Value

Returns a 3-row matrix. The columns are:

Estimate - bias adjusted standardized mean difference

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2020). “Point-biserial correlation: Interval estimation, hypothesis testing, meta-analysis, and sample size determination.” British Journal of Mathematical and Statistical Psychology, 73(S1), 113--144. ISSN 0007-1102, doi: 10.1111/bmsp.12189 , https://onlinelibrary.wiley.com/doi/10.1111/bmsp.12189.

Examples

ci.stdmean.strat(.05, 30.2, 30.8, 10.5, 11.2, 200, 200, .533)

#>                           Estimate         SE         LL        UL
#> Weighted standardizer: -0.05528428 0.10023259 -0.2518410 0.1410636
#> Group 1 standardizer:  -0.05692722 0.10368609 -0.2603639 0.1460782
#> Group 2 standardizer:  -0.05692722 0.09720571 -0.2440911 0.1369483

# Should return:
#                           Estimate         SE         LL        UL
# Weighted standardizer: -0.05528428 0.10023259 -0.2518410 0.1410636
# Group 1 standardizer:  -0.05692722 0.10368609 -0.2603639 0.1460782
# Group 2 standardizer:  -0.05692722 0.09720571 -0.2440911 0.1369483