Tukey-Kramer confidence intervals for all pairwise mean differences in a betwen-subjects design

Computes heteroscedastic Tukey-Kramer (also known as Games-Howell) confidence intervals for all pairwise comparisons of population means using sample means, sample standard deviations, and samples sizes as input. A Satterthwaite adjustment to the degrees of freedom is used to improve the accuracy of the confidence intervals.

ci.tukey(alpha, m, sd, n)

Arguments

alpha
alpha level for simultaneous 1-alpha confidence

m
vector of group sample means

sd
vector of group sample standard deviations

n
vector of sample sizes

Value

Returns a matrix with the number of rows equal to the number of pairwise comparisons. The columns are:

Estimate - estimated mean difference

SE - standard error

t - t test statistic

df - degrees of freedom

p - p-value

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Games PA, Howell JF (1976). “Pairwise multiple comparison procedures with unequal N's and/or variances: A Monte Carlo study.” Journal of Educational Statistics, 1(2), 113. ISSN 03629791, doi: 10.2307/1164979 , https://www.jstor.org/stable/1164979?origin=crossref.

Examples

m <- c(12.86, 17.57, 26.29)
sd <- c(3.185, 3.995, 3.773)
n <- c(20, 20, 20)
ci.tukey(.05, m, sd, n)

#>        diff       SE          t       df     p         LL         UL
#>  1 2  -4.71 1.142459  -4.122686 36.20303 0.001  -7.501842  -1.918158
#>  1 3 -13.43 1.104078 -12.163998 36.95915 0.000 -16.125713 -10.734287
#>  2 3  -8.72 1.228730  -7.096758 37.87646 0.000 -11.717053  -5.722947

# Should return:
#      Estimate       SE          t     p       df         LL         UL
# 1 2     -4.71 1.142459  -4.122686 0.001 36.20303  -7.501842  -1.918158
# 1 3    -13.43 1.104078 -12.163998 0.000 36.95915 -16.125713 -10.734287
# 2 3     -8.72 1.228730  -7.096758 0.000 37.87646 -11.717053  -5.722947